+++ src/Components/App/App.js
@@ -1,26 +1,27 @@
 import './App.css';
 import React, { useEffect } from 'react';
-import { getAllWorldsData } from './apiCalls'; 
+import { getAllWorldsData } from '../../apiCalls.js'
 import { useDispatch } from 'react-redux';
-import { getDiscoveredWorlds } from './Components/rootSlice';
+import { getDiscoveredWorlds } from '../rootSlice';
 import { Route, Switch } from 'react-router-dom';
-import { SingleWorld } from './Components/SingleWorld/SingleWorld';
-import { WelcomePage } from './Components/WelcomePage/WelcomePage'
-import { Header } from './Components/Header/Header.js'
-import { WorldBrowser } from './Components/WorldBrowser/WorldBrowser';
+import { SingleWorld } from '../SingleWorld/SingleWorld';
+import { WelcomePage } from '../WelcomePage/WelcomePage'
+import { Header } from '../Header/Header.js'
+import { WorldBrowser } from '../WorldBrowser/WorldBrowser';
 function App() {
   const dispatch = useDispatch()
 
   useEffect( () => {
     getAllWorldsData()
     .then(data => {
-      dispatch(getDiscoveredWorlds(data.worlds))
+      console.log(data)
+      dispatch(getDiscoveredWorlds(data))
     }).catch(err => console.log(err))
   }, [dispatch] )
 
   return (
     <div className="App">
-      <Header />
+      <Header/>
       <Switch>
         <Route path="/world/:id" render={() => ( <SingleWorld /> )}/>
         <Route path="/worlds" render={() => ( <WorldBrowser /> )} />
