+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './App';
-import { store } from './app/store';
+import App from './Components/App/App.js';
+import { store } from './Components/store';
 import { Provider } from 'react-redux'; 
 import { BrowserRouter } from 'react-router-dom';
